requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2.6
django-environ==0.4.5
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
openapi-codec==1.3.2
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
